+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-fix bug about scrape fans and share post
+1. 支持Instagram养号(随机点赞,关注,收藏,评论)和账号匹配
+2. Support Instagram follow and account match